Reuse connections to clamd daemon
Use clamd sessions to allow clamd connections be reused.
A connections pool is used inside clamd_mod where the connections stored after
use to be reused in future. This pool checked periodically and the expired
connections/sessions are closed.

Also the user can configure a maximum connections number to the clamd daemon.
If this number is reached, then the clamd_mod waits a connection to be free or
closed, before send a new request to clamd.

TODO: The clamd sessions allow queue multiple scan requests. Currently 
clamd_mod does not use this feature, but waits the active scan to be finished
before send a request for a new one.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap-modules@1347 950cb5d8-81e4-40ab-8cb8-c719d5138fe1