Fix: "Bug: service 'srv_clamav'. can not accept or sent body data"
This bug is hit when the client sent request with zero sized ICAP
preview data and the HTTP body data are more than the maximum
scannable object size. In this case the virus_scan service will
not unlock body data (will not allow sending body data, releasing
some space) and after some time will not be able to  store more
body data because there is not space. The service will enter a
state where it can not sent or receive body data for this request.