clamd_mod.c: Avoid "directive output may be truncated" compile warnings
While uses snprintf to generate CLAMD_VERSION and CLAMD_SIGNATURE
strings. This patch makes these strings longer.

Originally clamd_signature function which is an implementation of
av_engine::signature method has to return a string of size
SERVICE_ISTAG_SIZE, because this string was used directly to
ISTag header which has a specific size. Newer releases uses the
md5sum of this string so it can have any size we want